Move the text bloc data class to a separate file
